fix(gitrdun): write results file once per run, not twice
main() called display_results twice (once during the TUI, once after
it exited) and each call wrote a timestamped auto-filename file. The
two writes landed in different seconds, producing two output files
~1s apart with identical content.

Split formatting from I/O: format_results now returns the output
buffer as a pure function (no file-system side effects), and main
calls it exactly once. After the TUI exits, the captured buffer is
printed to stdout and written to disk exactly once via the new
write_output_file helper. This also removes the previously wasteful
second Ollama run.
